+++ src/main/java/bob/Storage.java
@@ -4,7 +4,7 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
-import bob.tasks.*;
+import bob.task.*;
 
 /**
  * Handles loading tasks from, and saving tasks to file
