+++ src/main/java/bob/command/AddCommand.java
@@ -1,4 +1,4 @@
-package bob.commands;
+package bob.command;
 import java.io.IOException;
 import bob.*;
