+++ src/main/java/bob/command/DeleteCommand.java
@@ -1,31 +1,27 @@
-package bob.commands;
+package bob.command;
 import java.io.IOException;
 import bob.*;
-import bob.exceptions.*;
+import bob.exception.*;
 
-public class MarkCommand extends Command {
-    public MarkCommand(String input) {
+public class DeleteCommand extends Command {
+    public DeleteCommand(String input) {
         super.input = input;
         super.isExit = false;
     }
 
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         try {
-            String[] markIndex = input.split(" ");
-            if (markIndex.length == 1) {
+            String[] deleteIndex = input.split(" ");
+            if (deleteIndex.length == 1) {
                 throw new MissingIndexException();
             }
             int index = 0;
             try {
-                index = Integer.parseInt(markIndex[1]);
+                index = Integer.parseInt(deleteIndex[1]);
             } catch (NumberFormatException e) {
                 throw new MissingIndexException();
             }
-            boolean doneOrNot = true;
-            if (input.contains("unmark")) {
-                doneOrNot = false;
-            }
-            ui.stringFormat(tasks.markDoneOrNot(index, doneOrNot));
+            ui.stringFormat(tasks.deleteTask(index));
             storage.write(tasks.lst);
         } catch (IOException e) {
             ui.showLoadingError();
