+++ src/main/java/bob/command/DisplayCommand.java
@@ -1,4 +1,4 @@
-package bob.commands;
+package bob.command;
 import bob.*;
 
 public class DisplayCommand extends Command {
