+++ src/main/java/bob/exception/BobException.java
@@ -1,4 +1,4 @@
-package bob.exceptions;
+package bob.exception;
 
 public class BobException extends Exception {
     public String message = "Error.";
