+++ src/main/java/bob/exception/MissingDeadlineException.java
@@ -1,4 +1,4 @@
-package bob.exceptions;
+package bob.exception;
 
 public class MissingDeadlineException extends BobException {
     public String message = "A due date must be given for deadlines!";
