+++ src/main/java/bob/exception/MissingEventDatesException.java
@@ -1,4 +1,4 @@
-package bob.exceptions;
+package bob.exception;
 
 public class MissingEventDatesException extends BobException {
     public String message = "Start and end times must be provided for events!";
