+++ src/main/java/bob/exception/MissingIndexException.java
@@ -1,4 +1,4 @@
-package bob.exceptions;
+package bob.exception;
 
 public class MissingIndexException extends BobException {
     public String message = "Task index must be provided!";
