+++ src/main/java/bob/exception/MissingTaskException.java
@@ -1,4 +1,4 @@
-package bob.exceptions;
+package bob.exception;
 
 public class MissingTaskException extends BobException {
     public String message = "Task description is missing!";
