+++ src/main/java/bob/exception/WrongInputException.java
@@ -1,4 +1,4 @@
-package bob.exceptions;
+package bob.exception;
 
 public class WrongInputException extends BobException {
     public String message = "Oh no, I don't understand that...";
