+++ src/main/java/bob/task/Deadline.java
@@ -1,7 +1,7 @@
-package bob.tasks;
+package bob.task;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import bob.exceptions.*;
+import bob.exception.*;
 
 public class Deadline extends Task {
     private LocalDate due;
