+++ src/main/java/bob/task/TaskType.java
@@ -1,4 +1,4 @@
-package bob.tasks;
+package bob.task;
 
 public enum TaskType {
     todo, event, deadline
